+++ Proyect/AutoLoads/Globals.gd
@@ -1,0 +1,4 @@
+extends Node
+
+var game
+var score_board