+++ changelog.md
@@ -1,2 +1,3 @@
-+ Goal score added.
-+ Working in NPC IA.
++ Adding Score Board Display.
++ Adding Mapping for GameControllers (Player 1).
++ Adding Roboto Font Asset.